requirement.txt
click
requests
logging
python-dotenv 
